Some more splitting of sbt-jni code.
