+++ .gitignore
@@ -23,3 +23,6 @@
 
 # Virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+# JNI configure output
+config.log
