+++ projects/build-basic/src/main/scala/sbt/tox4j/logic/jni/Platform.scala
@@ -1,0 +1,63 @@
+package sbt.tox4j.logic.jni
+
+import java.io.File
+
+import sbt.Keys._
+import sbt._
+import sbt.tox4j.Jni.Keys._
+
+sealed trait Platform
+
+object Platform {
+
+  final case class Android(platform: String) extends Platform
+  case object Host extends Platform
+
+  // Android build:
+  private def androidSettings(platform: String) = Seq(
+    toolchainPath := {
+      val candidates = Seq(
+        baseDirectory.value / "android" / platform,
+        baseDirectory.value.getParentFile / "android" / platform
+      )
+      // Try $TOOLCHAIN first, then try some other possible candidate paths.
+      sys.env.get("TOOLCHAIN").map(file).orElse(candidates.find(_.exists))
+    },
+    pkgConfigPath := toolchainPath.value.map(_ / "sysroot/usr/lib/pkgconfig").toSeq,
+    cppFlags := Nil,
+    ldFlags := Nil,
+
+    jniSourceFiles in Compile += {
+      def ifExists(file: File): Option[File] = {
+        Option(file).filter(_.exists)
+      }
+
+      val home = file(sys.env("HOME"))
+
+      val candidates = Seq(
+        sys.env.get("ANDROID_NDK_HOME").map(file).flatMap(ifExists),
+        ifExists(home / "usr/android-ndk"),
+        ifExists(home / "android-ndk")
+      )
+
+      candidates.find(_.nonEmpty).flatten match {
+        case Some(ndkHome) =>
+          val cpufeatures = ndkHome / "sources/android/cpufeatures/cpu-features.c"
+          if (!cpufeatures.exists) {
+            sys.error("Could not find cpu-features.c required for the Android build")
+          }
+          cpufeatures
+        case None =>
+          sys.error("Could not find Android NDK (you may need to set the ANDROID_NDK_HOME env var)")
+      }
+    }
+  )
+
+  def jniSettings(target: Platform): Seq[Setting[_]] = {
+    target match {
+      case Android(platform) => androidSettings(platform)
+      case Host              => Nil
+    }
+  }
+
+}
