+++ projects/tox4j/android/local.arm-linux-androideabi.sbt
@@ -1,0 +1,1 @@
+sbt.tox4j.logic.jni.Platform.Android("arm-linux-androideabi")
