+++ projects/tox4j/android/local.i686-linux-android.sbt
@@ -1,0 +1,1 @@
+sbt.tox4j.logic.jni.Platform.Android("i686-linux-android")
